requirements.txt
newsapi-python==0.2.6
clickhouse-driver==0.2.3
lz4==4.0.0
clickhouse-cityhash==1.0.2.3